requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-filter==2.3.0
djangorestframework-simplejwt==4.4.0
python-dotenv==0.14.0
pytest==5.4.1
pytest-django==3.9.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
